Moved from int to long for Twitter user IDs, adjusting to Twitter's move from 32-bit to 64-bit user IDs

+++ twitlogic-core/src/main/java/net/fortytwo/twitlogic/persistence/UserRegistry.java
@@ -37,7 +37,7 @@ public class UserRegistry {
         }
     }
 
-    public Integer resolveUserId(final String screenName) throws TwitterClientException {
+    public Long resolveUserId(final String screenName) throws TwitterClientException {
         User user = usersByScreenName.get(screenName);
 
         if (null == user) {
